New Struct, removed all parse history, created schemas for every collection, renamed files.
