+++ .gitignore
@@ -1,3 +1,3 @@
 
 .idea/workspace.xml
-WebApp/cloud/node_modules
+node_modules
