--- WebApp/.parse.local
@@ -1,10 +1,0 @@
-{
-  "applications": {
-    "Campus Connection": {
-      "applicationId": "tNMXiePfqLbNZP7ch3IES1OQ2QFU7SPGghtIH2ix"
-    },
-    "_default": {
-      "link": "Campus Connection"
-    }
-  }
-}
