--- WebApp/.parse.project
@@ -1,7 +1,0 @@
-{
-  "project_type": 1,
-  "parse": {
-    "jssdk": "latest"
-  },
-  "email": "johnfranklin42@gmail.com"
-}
