--- WebApp/cloud/models/sale.js
@@ -1,9 +1,0 @@
-/**
- * Created by johnfranklin on 2/19/16.
- */
-var db = require('../database');
-var sale = db.model('Sale', {
-    createdAt: {type: Date, required: true, default: Date.now}
-})
-
-module.exports = sale;
