--- WebApp/cloud/models/tag.js
@@ -1,9 +1,0 @@
-/**
- * Created by johnfranklin on 2/19/16.
- */
-var db = require('../database');
-var tag = db.model('Tag', {
-    createdAt: {type: Date, required: true, default: Date.now}
-})
-
-module.exports = tag;
