--- WebApp/cloud/models/university.js
@@ -1,8 +1,0 @@
-var db = require('../database')
-
-var University = db.model('University', {
-    name: String,
-    createdAt: {type: Date, required: true, default: Date.now}
-})
-
-module.exports = University
