+++ package.json
@@ -2,6 +2,7 @@
   "name": "application-name",
   "version": "0.0.1",
   "dependencies": {
+    "bower": "^1.7.7",
     "cookie-parser": "^1.4.1",
     "crypto": "0.0.3",
     "ejs": "^2.4.1",
