+++ public/components/LogOut/LogOut.js
