+++ public/components/viewListing/viewListing.js
