+++ server/models/product.js
@@ -1,0 +1,9 @@
+
+var mongoose = require('mongoose');
+
+var ProductSchema = new mongoose.Schema({
+    name: String,
+    createdAt: {type: Date, required: true, default: Date.now}
+});
+
+module.exports =  mongoose.model('Product', ProductSchema);
