+++ server/routes/imageRoutes.js
