+++ server/routes/userRoutes.js
@@ -3,8 +3,8 @@ var mongoose = require('mongoose');
 var passport = require('passport');
 var jwt = require('express-jwt');
 var auth = jwt({secret: 'SECRET', userProperty: 'payload'});
-require("../models/User");
-var User = mongoose.model('User');
+
+var User = require("../models/User");
 
 var router = express.Router();
